+++ ui/packages/shared/profile/src/ProfileIcicleGraph/index.tsx
@@ -15,8 +15,8 @@ import {useEffect, useMemo} from 'react';
 
 import {Flamegraph} from '@parca/client';
 import {Button} from '@parca/components';
-import {useContainerDimensions} from '@parca/dynamicsize';
-import {selectQueryParam, type NavigateFunction} from '@parca/functions';
+import {useContainerDimensions} from '@parca/hooks';
+import {selectQueryParam, type NavigateFunction} from '@parca/utilities';
 
 import DiffLegend from '../components/DiffLegend';
 import {IcicleGraph} from './IcicleGraph';
@@ -28,6 +28,8 @@ export type ResizeHandler = (width: number, height: number) => void;
 interface ProfileIcicleGraphProps {
   width?: number;
   graph: Flamegraph | undefined;
+  total: bigint;
+  filtered: bigint;
   sampleUnit: string;
   curPath: string[] | [];
   setNewCurPath: (path: string[]) => void;
@@ -39,6 +41,8 @@ interface ProfileIcicleGraphProps {
 
 const ProfileIcicleGraph = ({
   graph,
+  total,
+  filtered,
   curPath,
   setNewCurPath,
   sampleUnit,
@@ -58,25 +62,36 @@ const ProfileIcicleGraph = ({
     onContainerResize(dimensions.width, dimensions.height);
   }, [dimensions, onContainerResize]);
 
-  const [isTrimmed, _, trimmedPercentage, formattedTotal, formattedUntrimmedTotal] = useMemo(() => {
-    if (graph === undefined || graph.untrimmedTotal === '0') {
-      return [false, BigInt(0), '0'];
+  const [
+    totalFormatted,
+    rawFormatted,
+    isTrimmed,
+    trimmedFormatted,
+    trimmedPercentage,
+    isFiltered,
+    filteredPercentage,
+  ] = useMemo(() => {
+    if (graph === undefined) {
+      return ['0', '0', false, '0', '0', false, '0', '0'];
     }
 
-    const untrimmedTotal = BigInt(graph.untrimmedTotal);
-    const total = BigInt(graph.total);
+    const trimmed = BigInt(graph.trimmed);
 
-    const trimDifference = untrimmedTotal - total;
-    const trimmedPercentage = (total * BigInt(100)) / untrimmedTotal;
+    const rawTotal = total + filtered + trimmed;
+
+    // safeguard against division by zero
+    const rawTotalDivisor = rawTotal > 0 ? rawTotal : BigInt(1);
 
     return [
-      trimDifference > BigInt(0),
-      trimDifference,
-      trimmedPercentage.toString(),
       numberFormatter.format(total),
-      numberFormatter.format(untrimmedTotal),
+      numberFormatter.format(rawTotal),
+      trimmed > 0,
+      numberFormatter.format(trimmed),
+      numberFormatter.format((trimmed * BigInt(100)) / rawTotalDivisor),
+      filtered > 0,
+      numberFormatter.format((total * BigInt(100)) / rawTotalDivisor),
     ];
-  }, [graph]);
+  }, [filtered, graph, total]);
 
   useEffect(() => {
     if (setActionButtons === undefined) {
@@ -98,18 +113,15 @@ const ProfileIcicleGraph = ({
 
   if (graph === undefined) return <div>no data...</div>;
 
-  const total = graph.total;
+  if (total === BigInt(0) && !loading) return <>Profile has no samples</>;
 
-  if (parseFloat(total) === 0 && !loading) return <>Profile has no samples</>;
+  if (isTrimmed) {
+    console.info(`Trimmed ${trimmedFormatted} (${trimmedPercentage}%) too small values.`);
+  }
 
   return (
     <div className="relative">
       {compareMode && <DiffLegend />}
-      {isTrimmed ? (
-        <p className="my-2 text-sm">
-          Showing {formattedTotal}({trimmedPercentage}%) out of {formattedUntrimmedTotal} samples
-        </p>
-      ) : null}
       <div ref={ref}>
         <IcicleGraph
           width={dimensions?.width}
@@ -120,6 +132,17 @@ const ProfileIcicleGraph = ({
           navigateTo={navigateTo}
         />
       </div>
+      <p className="my-2 text-xs">
+        Showing {totalFormatted}{' '}
+        {isFiltered ? (
+          <span>
+            ({filteredPercentage}%) filtered of {rawFormatted}{' '}
+          </span>
+        ) : (
+          <></>
+        )}
+        values.{' '}
+      </p>
     </div>
   );
 };